refactor. 예약 완료 스케쥴러 분리
