+++ BE/src/main/java/team07/airbnb/service/booking/BookingInquiryService.java
@@ -1,14 +1,12 @@
 package team07.airbnb.service.booking;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import team07.airbnb.controller.ReviewController;
 import team07.airbnb.data.booking.dto.response.BookingDetailResponse;
 import team07.airbnb.data.booking.dto.transfer.DateInfo;
 import team07.airbnb.data.booking.dto.transfer.DistanceInfo;
-import team07.airbnb.data.booking.enums.BookingStatus;
 import team07.airbnb.data.review.dto.request.ReviewPostRequest;
 import team07.airbnb.entity.AccommodationEntity;
 import team07.airbnb.entity.BookingEntity;
@@ -21,7 +19,6 @@ import team07.airbnb.repository.BookingRepository;
 import team07.airbnb.service.accommodation.AccommodationService;
 import team07.airbnb.service.product.ProductService;
 
-import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -31,23 +28,7 @@ public class BookingInquiryService {
 
     private final AccommodationService accommodationService;
     private final ProductService productService;
-
     private final BookingRepository bookingRepository;
-
-    /**
-     * 정오마다 오늘이 체크아웃 날짜인 모든 예약을 이용 완료 상태로 변경
-     */
-    @Scheduled(cron = "0 0 12 * * ?")
-    public void completeAllBookings() {
-        LocalDate now = LocalDate.now();
-
-        List<BookingEntity> bookings = bookingRepository.findAllByCheckout(now);
-
-        bookings.forEach(booking -> {
-            booking.setStatus(BookingStatus.COMPLETE);
-            booking.getProducts().stream().close();
-        });
-    }
 
 
     public List<BookingDetailResponse> findBookingsByUser(UserEntity booker) {
